+++ src/remotes/dzrpbuffer/cspectremote.ts
@@ -33,11 +33,6 @@ export class CSpectRemote extends DzrpBufferRemote {
 	/// The successful emit takes place in 'onConnect' which should be called
 	/// by 'doInitialization' after a successful connect.
 	public async doInitialization(): Promise<void>  {
-		// Check for unsupported settings
-		if (Settings.launch.unitTests) {
-			throw Error("launch.json: unitTests==true: CSpect does not support running unit tests.");
-		}
-
 		// Init socket
 		this.socket=new Socket();
 		this.socket.unref();